refactor(arch): Phase 10 — PR #490 리뷰 수정
- B1/B2: periodLabels, marketHoursDisplay → shared/lib/ (presentation 상수)
- optionsFormatters barrel export 추가 (formatAtmIv 포함)
- widget import → barrel/shared 경로로 갱신
