+++ src/shared/lib/news/periodLabels.ts
