+++ src/widgets/options/OptionsStaleDataBanner.tsx
@@ -4,7 +4,7 @@ import {
     ET_MARKET_HOURS_DISPLAY,
     KST_EDT_HOURS_DISPLAY,
     KST_EST_HOURS_DISPLAY,
-} from '@/entities/options-chain/lib/marketHoursDisplay';
+} from '@/shared/lib/options/marketHoursDisplay';
 import { EDT_OFFSET_HOURS, getEasternOffsetHours } from '@/shared/lib/eastern';
 
 /**
