+++ src/widgets/options/utils/optionsTooltips.tsx
@@ -25,7 +25,7 @@ import {
     ET_MARKET_HOURS_DISPLAY,
     KST_EDT_HOURS_DISPLAY,
     KST_EST_HOURS_DISPLAY,
-} from '@/entities/options-chain/lib/marketHoursDisplay';
+} from '@/shared/lib/options/marketHoursDisplay';
 
 interface KstWindowInfo {
     window: string;
